Add int precision

+++ static_lib/add_numbers.c
@@ -1,4 +1,4 @@
-int add_numbers (long int x, long int y)
+int add_numbers (long long x, long long y)
 {
     return x + y;
 }
